Add 'Run Rust tests' Github action
This uses Github actions to automatically run the Rust tests on all three
major platforms.
This triggers when doing a pull request on the 'master' branch, or when
manually launched via the 'Actions' panel on Github.

Note that I couldn't make the tests work on windows, as `gdnative`'s
compilation fails with `Unable to find libclang: "the `libclang` shared
library at C:\\msys64\\mingw64\\bin\\libclang.dll could not be opened:
LoadLibraryExW failed"`.