Add a 'Build shared libraries' Github action
This action can be triggered manually via the 'Action' panel on Github.
It will then build the shared libraries for linux, windows and macos and put
them in the correct Dijkstra_map_library/bin/<os> repository.
It will generate both a new commit and a tag.

Note: I couldn't make the build on windows, but building it on linux with the
x86_64-pc-windows-gnu target worked (all details are in the
.github/workflows/make_release.yml file).

I also tested all 3 library files on their respective machines, and it seems
to work without issue.