Do not manually update binaries
This adds a test that forbid binaries for being updated by a pull request. Binaries should be updated using the corresponding github action.

This ensures that all 3 binaries are always at the same release, and also that they are built predictably. In particular, this should allow us to pinpoint a specific glibc version on linux.